+++ src/o_init.c
@@ -251,7 +251,7 @@ savenames(fd, mode)
 int fd, mode;
 {
 	register int i;
-	unsigned int len;
+	size_t len;
 
 	if (perform_bwrite(mode)) {
 	    bwrite(fd, (genericptr_t)bases, sizeof bases);
