+++ src/version.c
@@ -105,7 +105,7 @@ uptodate(fd, name)
 int fd;
 const char *name;
 {
-    int rlen;
+    ssize_t rlen;
     struct version_info vers_info;
     boolean verbose = name ? TRUE : FALSE;
